Remove integration tests for Swin_s and Enable weights from CIv2
